iCalendar Reader: Set a BYDAY for WEEKLY events.
We treat BYDAY specially while parsing iCal files to deal with recurring event timestamps. An event that takes place with a FREQ=WEEKLY will always have a BYDAY, even if not specified in the iCal file.

This sets the BYDAY for a WEEKLY events to avoid falling back on more generic event-parsing code.

This fixes an issue with a processing timestamp for a WEEKLY event without a BYDAY; the timestamp of the displayed event incremented with each occurrence by the GMT offset.

See https://[private link]

Merges r117277-wpcom.